[iOS] [12538] Remove tabs from `Events` section
